Core Update
- Migrate Webpack to TypeScript
- Remove plugins
- Removed playground
- Migrated some parts to react-onsenuix
- Perfo, improvements
