+++ .prettierrc
@@ -1,5 +1,5 @@
 {
-  "printWidth": 500,
+  "printWidth": 200,
   "tabWidth": 2,
   "useTabs": false,
   "semi": true,
