+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "hentai-web",
-  "version": "1.4.1",
+  "version": "1.4.2",
   "description": "",
   "main": "src/index.tsx",
   "keywords": [
@@ -28,17 +28,20 @@
   "author": "Der_Googler",
   "license": "ISC",
   "scripts": {
-    "build-dev": " webpack --mode development",
-    "build-prod": "webpack --mode production",
+    "dev": "webpack --config webpack.dev.ts",
+    "prod": "bun run dev && webpack --config webpack.prod.ts && rm ./dist/**/*.map",
     "make-types": "npx -p typescript tsc src/**/*.ts --declaration --outDir dist/types",
     "start": "webpack-dev-server"
+  },
+  "resolutions": {
+    "@types/react": "18.0.15"
   },
   "dependencies": {
     "@babel/polyfill": "^7.12.1",
     "@babel/standalone": "^7.17.6",
     "@material-ui/core": "^4.12.3",
     "@monaco-editor/react": "^4.3.1",
-    "@types/react": "^17.0.32",
+    "@types/react": "^18.0.15",
     "@types/react-onsenui": "^2.9.16",
     "ace-builds": "^1.4.14",
     "autoprefixer": "^10.2.5",
@@ -49,12 +52,12 @@
     "boolean": "^3.1.4",
     "bootstrap": "^5.1.3",
     "bootstrap-icons": "^1.7.2",
-    "bota64": "github:DerGoogler/Bota64",
     "core-js": "^3.18.1",
     "crypto-js": "^4.1.1",
     "eruda": "^2.4.1",
     "eruda-dom": "^2.0.0",
     "file-saver": "^2.0.5",
+    "googlers-tools": "^1.2.5",
     "highlight.js": "^11.4.0",
     "hmfull": "^1.3.0",
     "js-yaml": "^4.1.0",
@@ -71,12 +74,13 @@
     "node-eval": "^2.0.0",
     "onsenui": "^2.11.2",
     "path-browserify": "^1.0.1",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-ace": "^9.5.0",
     "react-bootstrap": "^2.0.3",
     "react-device-detect": "^2.1.2",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-onsenui": "^1.11.5",
+    "react-onsenuix": "^1.0.5",
     "react-translated": "^2.5.0",
     "safer-eval": "^1.3.6",
     "universal-cookie": "^4.0.4",
@@ -118,9 +122,10 @@
     "ts-loader": "^9.2.6",
     "ts-node": "^10.4.0",
     "typescript": "^4.4.4",
+    "uglifyjs-webpack-plugin": "^2.2.0",
     "url-loader": "^4.1.1",
     "vscode-theme-to-monaco-theme-node": "^1.0.0",
-    "webpack": "^5.65.0",
+    "webpack": "^5.73.0",
     "webpack-cli": "^4.9.2",
     "webpack-sources": "^3.2.3"
   }
