+++ src/components/Gesture.tsx
@@ -1,38 +1,41 @@
 import tools from "@Misc/tools";
 import * as React from "react";
+import { ViewX, ViewXRenderData } from "react-onsenuix";
 
-class Gesture extends React.Component<
-  {
-    event:
-      | "drag"
-      | "dragleft"
-      | "dragright"
-      | "dragup"
-      | "dragdown"
-      | "hold"
-      | "release"
-      | "swipe"
-      | "swipeleft"
-      | "swiperight"
-      | "swipeup"
-      | "swipedown"
-      | "tap"
-      | "doubletap"
-      | "touch"
-      | "transform"
-      | "pinch"
-      | "pinchin"
-      | "pinchout"
-      | "rotate";
-    callback(...props: any): void;
-  },
-  {}
-> {
+interface Props {
+  event:
+    | "drag"
+    | "dragleft"
+    | "dragright"
+    | "dragup"
+    | "dragdown"
+    | "hold"
+    | "release"
+    | "swipe"
+    | "swipeleft"
+    | "swiperight"
+    | "swipeup"
+    | "swipedown"
+    | "tap"
+    | "doubletap"
+    | "touch"
+    | "transform"
+    | "pinch"
+    | "pinchin"
+    | "pinchout"
+    | "rotate";
+  callback(...props: any): void;
+}
+
+class Gesture extends ViewX<Props, {}> {
   private gerstureID: React.RefObject<HTMLDivElement>;
-  constructor(props: any) {
+
+  public constructor(props: any) {
     super(props);
     this.gerstureID = React.createRef();
+    this.createView = this.createView.bind(this);
   }
+
   public componentDidMount() {
     const { callback, event } = this.props;
 
@@ -41,9 +44,8 @@ class Gesture extends React.Component<
     });
   }
 
-  public render() {
-    const { children } = this.props;
-    return <div ref={this.gerstureID}>{children}</div>;
+  public createView(data: ViewXRenderData<Props, {}, HTMLElement>): JSX.Element {
+    return <div ref={this.gerstureID}>{data.p.children}</div>;
   }
 }
