--- src/views/EditorActivity/interface.ts
@@ -1,11 +1,0 @@
-interface Props {
-  extras: any;
-  popPage: any;
-}
-interface States {
-  isContextOpen: boolean;
-  data: string;
-  types: any;
-}
-
-export { Props, States };
