+++ src/views/index.ts
@@ -1,23 +1,15 @@
 import InitActivity from "./InitActivity";
 import MainActivity from "./MainActivity";
 import ChangelogActivity from "./ChangelogActvity";
-import PluginAboutActivity from "./plugin/PluginAboutActivity";
 import SettingsActivity from "./SettingsActivity";
-import EditorActivity from "./EditorActivity";
-import ForbiddenActivity from "./ForbiddenActivity";
 import TextFetchActivity from "./TextFetchActivity";
 import LoginActivity from "./LoginActivity";
-import PluginsActivity from "./plugin/PluginsActvity";
 
 export {
   InitActivity,
-  PluginsActivity,
   MainActivity,
   LoginActivity,
   ChangelogActivity,
-  PluginAboutActivity,
   SettingsActivity,
-  EditorActivity,
-  ForbiddenActivity,
   TextFetchActivity,
 };
