--- src/views/plugin/PluginAboutActivity/interface.ts
@@ -1,11 +1,0 @@
-import { PushPageProps } from "@Types/init";
-
-interface Props {
-  pushPage(props: PushPageProps): void;
-  popPage: any;
-  pluginAbout: any;
-}
-
-interface States {}
-
-export { Props, States };
