--- src/views/plugin/PluginsActvity/interface.ts
@@ -1,10 +1,0 @@
-import { PushPageProps } from "@Types/init";
-
-interface Props {
-  pushPage(props: PushPageProps): void;
-  popPage: any;
-}
-
-interface States {}
-
-export { Props, States };
